refactor(project): move REST handlers to module
