+++ src/server/routes/mod.rs
@@ -1,5 +1,5 @@
 //! Route configuration and handlers.
 pub mod health;
 pub mod inventory;
-pub mod projects;
+pub mod project;
 pub mod router;
