+++ src/server/routes/project/mod.rs
@@ -1,0 +1,5 @@
+pub mod register;
+pub mod view;
+
+
+
