+++ src/server/routes/router.rs
@@ -5,7 +5,8 @@ use std::sync::Arc;
 use crate::inventory::app::service::InventoryService;
 use crate::projects::app::service::ProjectsService;
 use crate::server::routes::health::health;
-use crate::server::routes::projects::{register_project, view_project};
+use crate::server::routes::project::register::register_project;
+use crate::server::routes::project::view::view_project;
 use axum::routing::{get, post};
 use axum::Router;
 use tower::ServiceBuilder;
